create note to notebook

+++ src/com/android/mydm/remote/BackgroundService.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 import org.apache.thrift.TException;
 import org.apache.thrift.transport.TTransportException;
 
+import com.android.mydm.R;
 import com.android.mydm.method.CreateNote;
 import com.android.mydm.method.CreateNote.CreateNoteParams;
 import com.android.mydm.util.EvernoteUtil;
@@ -19,6 +20,7 @@ import android.app.IntentService;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
+import android.widget.Toast;
 
 public class BackgroundService extends IntentService {
 	public static final String ACTION_CREATE_NOTE = "com.android.text.ACTION_CREATE_NOTE";
@@ -49,20 +51,28 @@ public class BackgroundService extends IntentService {
 				params.title = extras.getString("title");
 				params.description = extras.getString("description");
 				params.mimeType = extras.getString("mimeType");
-				File file = new File(intent.getData().getPath());
-				Log.d(LOG_TAG, "file exists " + file.exists());
-				params.f = file.getAbsolutePath();
-				if (params.f.toLowerCase().endsWith("jpg")) {
-					params.mimeType = "image/jpeg";
-				} else {
-					params.mimeType = "image/png";
+				params.notebookId = extras.getString("notebookId");
+				File file = null;
+				if (intent.getData() != null) {
+					file = new File(intent.getData().getPath());
+					Log.d(LOG_TAG, "file exists " + file.exists());
+					params.f = file.getAbsolutePath();
+					if (params.f.toLowerCase().endsWith("jpg")) {
+						params.mimeType = "image/jpeg";
+					} else {
+						params.mimeType = "image/png";
+					}
 				}
 				String tag = intent.getStringExtra("tag");
 				params.tagNames = new ArrayList<String>();
 				params.tagNames.add(tag);
 				create.execute(params);
-				file.delete();
-				Log.d(LOG_TAG, "success");
+				if (file != null) {
+					file.delete();
+				}
+				Toast.makeText(getBaseContext(),
+						getString(R.string.create_success), Toast.LENGTH_SHORT)
+						.show();
 			}
 		} catch (TTransportException e) {
 			Log.e(LOG_TAG, "error", e);
